# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Production API for Shiprion quotes, shipments, tracking, accounts, support, and operations.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rates/estimate:
    post:
      operationId: estimateRate
      tags: [rates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceCode, weightKg]
              properties:
                serviceCode: { type: string }
                weightKg: { type: number, exclusiveMinimum: 0 }
      responses:
        "200":
          description: Server-calculated estimate
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RateEstimate" } },
            }
  /admin/rates:
    get:
      operationId: listAllRates
      tags: [rates]
      responses:
        "200": { description: Every service-rate version }
    post:
      operationId: createRateVersion
      tags: [rates]
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/CreateRateBody" } },
          }
      responses:
        "201": { description: Activated service-rate version }
  /admin/rates/{id}:
    patch:
      operationId: setRateActive
      tags: [rates]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [active],
                    properties: { active: { type: boolean } },
                  },
              },
          }
      responses:
        "200": { description: Updated rate activation }
...
components:
  schemas:
...
    ServiceRate:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: integer }
        baseFeeCents: { type: integer }
        perKgCents: { type: integer }
        fuelPct: { type: number }
        minWeightKg: { type: number }
        maxWeightKg: { type: number }
        transitDaysMin: { type: integer }
        transitDaysMax: { type: integer }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
      required:
        [
          id,
          code,
          name,
          description,
          version,
          baseFeeCents,
          perKgCents,
          fuelPct,
          minWeightKg,
          maxWeightKg,
          transitDaysMin,
          transitDaysMax,
          active,
          createdAt,
        ]
...
    PriceBreakdown:
      type: object
      properties:
        baseFeeCents: { type: integer }
        weightFeeCents: { type: integer }
        fuelSurchargeCents: { type: integer }
        totalCents: { type: integer }
        currency: { type: string, enum: [USD] }
      required:
        [baseFeeCents, weightFeeCents, fuelSurchargeCents, totalCents, currency]
    RateEstimate:
      type: object
      properties:
        rate: { $ref: "#/components/schemas/ServiceRate" }
        breakdown: { $ref: "#/components/schemas/PriceBreakdown" }
      required: [rate, breakdown]
    CreateRateBody:
      type: object
      properties:
        code: { type: string, pattern: "^[a-z0-9-]{2,40}$" }
        name: { type: string, minLength: 2, maxLength: 80 }
        description: { type: string, maxLength: 300 }
        baseFeeCents: { type: integer, minimum: 0 }
        perKgCents: { type: integer, minimum: 0 }
        fuelPct: { type: number, minimum: 0, maximum: 100 }
        minWeightKg: { type: number, exclusiveMinimum: 0 }
        maxWeightKg: { type: number, exclusiveMinimum: 0 }
        transitDaysMin: { type: integer, minimum: 0 }
        transitDaysMax: { type: integer, minimum: 0 }
      required:
        [
          code,
          name,
          description,
          baseFeeCents,
          perKgCents,
          fuelPct,
          minWeightKg,
          maxWeightKg,
          transitDaysMin,
          transitDaysMax,
        ]
...